+++ week3/Experiment3.java
@@ -4,7 +4,7 @@ public class Experiment3 {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
 
-        int price;
+        double price;
         double discount;
         double discountCode = 0.15;
         double totalPayment;
@@ -16,7 +16,7 @@ public class Experiment3 {
         totalPayment = price - discount;
 
         System.out.println("Original Price: Rp " + price);
-        System.out.println("Discount (15%): Rp " +discount);
+        System.out.println("Discount (15%): Rp " + discount);
         System.out.println("Total to be Paid: Rp " + totalPayment);
     }
 }
